+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - [changelog](https://github.com/getsentry/sentry-wizard/blob/master/CHANGELOG.md#140)
   - [diff](https://github.com/getsentry/sentry-wizard/compare/v1.2.17...v1.4.0)
 - Android builds without ext config, auto create assets dir for modules ([#2652](https://github.com/getsentry/sentry-react-native/pull/2652))
+- Exit gracefully if source map file for collecting modules doesn't exist ([#2655](https://github.com/getsentry/sentry-react-native/pull/2655))
 
 ### Dependencies